Move serialization v0 to a tests-only java file
